+++ server/routes/index.ts
@@ -34,7 +34,7 @@ import InstallationLocationController from '../controllers/monitoringConditions/
 import ReceiptController from '../controllers/receiptController'
 import AttachmentHavePhotoController from '../controllers/attachments/attachmentHavePhotoController'
 import IsRejectionController from './is-rejection/controller'
-import createOrderTypeDescriptionRouter from './monitoring-conditions/order-type/router'
+import createOrderTypeDescriptionRouter from './monitoring-conditions/router'
 
 export default function routes({
   alcoholMonitoringService,
